feat: 실습 피드백 | sections -> 일급 컬렉션
