+++ src/test/java/nextstep/subway/unit/LineServiceTest.java
@@ -46,7 +46,7 @@ public class LineServiceTest {
 
         // then
         // line.getSections 메서드를 통해 검증
-        final List<Section> sections = line.getSections();
-        assertThat(sections).isNotEmpty();
+        final Sections sections = line.sections();
+        assertThat(sections.isEmpty()).isFalse();
     }
 }
